Make sure produce_block doesn't produce blocks at the same height
