+++ chain/chunks/src/lib.rs
@@ -720,7 +720,7 @@ impl ShardsManager {
         receipts: &Vec<Receipt>,
         receipts_root: CryptoHash,
         tx_root: CryptoHash,
-        signer: Arc<dyn Signer>,
+        signer: &dyn Signer,
     ) -> Result<EncodedShardChunk, Error> {
         let total_parts = self.runtime_adapter.num_total_parts(&prev_block_hash);
         let data_parts = self.runtime_adapter.num_data_parts(&prev_block_hash);
