+++ chain/client/tests/challenges.rs
@@ -3,120 +3,61 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use near_chain::{Block, ChainGenesis, Provenance};
-use near_client::test_utils::{setup_client, MockNetworkAdapter};
-use near_client::Client;
+use near_client::test_utils::{setup_client, MockNetworkAdapter, TestEnv};
 use near_crypto::{InMemorySigner, KeyType};
 use near_network::types::{ChunkOnePartRequestMsg, PeerId};
-use near_network::NetworkClientResponses;
 use near_primitives::block::BlockHeader;
 use near_primitives::challenge::Challenge;
 use near_primitives::hash::CryptoHash;
 use near_primitives::serialize::BaseDecode;
 use near_primitives::sharding::EncodedShardChunk;
 use near_primitives::test_utils::init_test_logger;
-use near_primitives::transaction::SignedTransaction;
-use near_primitives::types::{AccountId, BlockIndex, MerkleHash};
+use near_primitives::types::MerkleHash;
 use near_store::test_utils::create_test_store;
 
-struct TestEnv {
-    chain_genesis: ChainGenesis,
-    validators: Vec<AccountId>,
-    pub network_adapters: Vec<Arc<MockNetworkAdapter>>,
-    pub clients: Vec<Client>,
-}
-
-impl TestEnv {
-    pub fn new(num_clients: usize, num_validators: usize) -> Self {
-        let chain_genesis = ChainGenesis::test();
-        let validators: Vec<AccountId> =
-            (0..num_validators).map(|i| format!("test{}", i)).collect();
-        let network_adapters =
-            (0..num_clients).map(|_| Arc::new(MockNetworkAdapter::default())).collect::<Vec<_>>();
-        let clients = (0..num_clients)
-            .map(|i| {
-                let store = create_test_store();
-                setup_client(
-                    store.clone(),
-                    vec![validators.iter().map(|x| x.as_str()).collect::<Vec<&str>>()],
-                    1,
-                    1,
-                    Some(&format!("test{}", i)),
-                    network_adapters[i].clone(),
-                    chain_genesis.clone(),
-                )
-            })
-            .collect();
-        TestEnv { chain_genesis, validators, network_adapters, clients }
-    }
-
-    pub fn produce_block(&mut self, id: usize, height: BlockIndex) {
-        let block = self.clients[id].produce_block(height, Duration::from_millis(10)).unwrap();
-        let (mut accepted_blocks, _) =
-            self.clients[id].process_block(block.clone().unwrap(), Provenance::PRODUCED);
-        let more_accepted_blocks = self.clients[id].run_catchup().unwrap();
-        accepted_blocks.extend(more_accepted_blocks);
-        for accepted_block in accepted_blocks.into_iter() {
-            self.clients[id].on_block_accepted(
-                accepted_block.hash,
-                accepted_block.status,
-                accepted_block.provenance,
-            );
-        }
-    }
-
-    pub fn send_money(&mut self, id: usize) -> NetworkClientResponses {
-        let signer = InMemorySigner::from_seed("test1", KeyType::ED25519, "test1");
-        let tx = SignedTransaction::send_money(
-            1,
-            "test1".to_string(),
-            "test1".to_string(),
-            &signer,
-            100,
-            self.clients[id].chain.head().unwrap().last_block_hash,
-        );
-        self.clients[id].process_tx(tx)
-    }
-
-    pub fn restart(&mut self, id: usize) {
-        let store = self.clients[id].chain.store().store().clone();
-        self.clients[id] = setup_client(
-            store,
-            vec![self.validators.iter().map(|x| x.as_str()).collect::<Vec<&str>>()],
-            1,
-            1,
-            Some(&format!("test{}", id)),
-            self.network_adapters[id].clone(),
-            self.chain_genesis.clone(),
-        )
-    }
-}
-
 #[test]
-fn test_verify_double_sign_challenge() {
-    let mut env = TestEnv::new(2, 1);
+fn test_verify_block_double_sign_challenge() {
+    let mut env = TestEnv::new(ChainGenesis::test(), 2, 1);
     env.produce_block(0, 1);
+    let genesis = env.clients[0].chain.get_block_by_height(0).unwrap().clone();
     let b1 = env.clients[0].produce_block(2, Duration::from_millis(10)).unwrap().unwrap();
-    let b2 = env.clients[0].produce_block(2, Duration::from_millis(10)).unwrap().unwrap();
+    let signer = InMemorySigner::from_seed("test0", KeyType::ED25519, "test0");
+    let b2 = Block::produce(
+        &genesis.header,
+        2,
+        genesis.chunks.clone(),
+        b1.header.inner.epoch_id.clone(),
+        HashMap::default(),
+        0,
+        None,
+        &signer,
+    );
     let valid_challenge = Challenge::BlockDoubleSign {
         left_block_header: b1.header.clone(),
         right_block_header: b2.header.clone(),
     };
-    assert!(env.clients[0].verify_challenge(valid_challenge).unwrap());
+    assert!(env.clients[1].verify_challenge(valid_challenge).unwrap());
     let invalid_challenge = Challenge::BlockDoubleSign {
         left_block_header: b1.header.clone(),
         right_block_header: b1.header.clone(),
     };
-    assert!(!env.clients[0].verify_challenge(invalid_challenge).unwrap());
+    assert!(!env.clients[1].verify_challenge(invalid_challenge).unwrap());
     let b3 = env.clients[0].produce_block(3, Duration::from_millis(10)).unwrap().unwrap();
     let invalid_challenge =
         Challenge::BlockDoubleSign { left_block_header: b1.header, right_block_header: b3.header };
-    assert!(!env.clients[0].verify_challenge(invalid_challenge).unwrap());
+    assert!(!env.clients[1].verify_challenge(invalid_challenge).unwrap());
+}
+
+#[test]
+fn test_verify_chunk_double_sign_challenge() {
+    let mut env = TestEnv::new(ChainGenesis::test(), 1, 1);
+    // env.clients[0].pro
 }
 
 #[test]
 fn test_request_chunk_restart() {
     init_test_logger();
-    let mut env = TestEnv::new(1, 1);
+    let mut env = TestEnv::new(ChainGenesis::test(), 1, 1);
     for i in 1..3 {
         env.produce_block(0, i);
         env.network_adapters[0].pop();
@@ -149,7 +90,7 @@ fn create_block_with_invalid_chunk(
     prev_block_header: &BlockHeader,
     account_id: &str,
 ) -> (Block, EncodedShardChunk) {
-    let signer = Arc::new(InMemorySigner::from_seed(account_id, KeyType::ED25519, account_id));
+    let signer = InMemorySigner::from_seed(account_id, KeyType::ED25519, account_id);
     let (invalid_encoded_chunk, _merkle_paths) = EncodedShardChunk::new(
         prev_block_header.hash,
         CryptoHash::from_base("F5SvmQcKqekuKPJgLUNFgjB4ZgVmmiHsbDhTBSQbiywf").unwrap(),
@@ -165,7 +106,7 @@ fn create_block_with_invalid_chunk(
         &vec![],
         &vec![],
         MerkleHash::default(),
-        signer.clone(),
+        &signer,
     )
     .unwrap();
     let block_with_invalid_chunk = Block::produce(
@@ -176,7 +117,7 @@ fn create_block_with_invalid_chunk(
         HashMap::default(),
         0,
         None,
-        signer,
+        &signer,
     );
     (block_with_invalid_chunk, invalid_encoded_chunk)
 }
