+++ core/primitives/src/test_utils.rs
@@ -1,5 +1,5 @@
 use std::collections::HashMap;
-use std::sync::{Arc, Mutex};
+use std::sync::Mutex;
 
 use log::LevelFilter;
 
@@ -134,16 +134,16 @@ impl Block {
         prev: &Block,
         height: BlockIndex,
         epoch_id: EpochId,
-        signer: Arc<dyn Signer>,
+        signer: &dyn Signer,
     ) -> Self {
         Self::empty_with_approvals(prev, height, epoch_id, HashMap::default(), signer)
     }
 
-    pub fn empty_with_height(prev: &Block, height: BlockIndex, signer: Arc<dyn Signer>) -> Self {
+    pub fn empty_with_height(prev: &Block, height: BlockIndex, signer: &dyn Signer) -> Self {
         Self::empty_with_epoch(prev, height, prev.header.inner.epoch_id.clone(), signer)
     }
 
-    pub fn empty(prev: &Block, signer: Arc<dyn Signer>) -> Self {
+    pub fn empty(prev: &Block, signer: &dyn Signer) -> Self {
         Self::empty_with_height(prev, prev.header.inner.height + 1, signer)
     }
 
@@ -154,7 +154,7 @@ impl Block {
         height: BlockIndex,
         epoch_id: EpochId,
         approvals: HashMap<usize, Signature>,
-        signer: Arc<dyn Signer>,
+        signer: &dyn Signer,
     ) -> Self {
         Block::produce(
             &prev.header,
